Move `fissile build images` logic into separate build_images.go
